0.3v
- Removed the Gift Incident,  just wasn't interesting.

- Made a window for a war, showing resorucee and war events history.
- Change War overview from comms to Factions tab through a button

-- Fixed a bug with the mercenary incident when a faction didn't have atleader. It will now generate a new one.

- Deleted unnecessery usings, casts and "this", compacted lines

- Changed history lines prefixes.

- War is now started through an incident instead of a comp.

- Trade vassal investment now gives a trade discount with the vassal faction.

- Fixed the Aid incident to correctly know when it can fire instead of returning false in the worker.

- Added a settings to turn off the destructive war outcomes.

- Changed the investments ints to bytes because why not.

- Text fixes

- Added faction resources boost for each investment in a vassal.

- Fixed an issue that every faction would expand at once for the first expansion.

- Fixed the mercenary battle incident: better pawn spawning, fixed spawning tile and fixed battle end detection.

- A sweep of all incidients and their comps to fix any problems found.

+++ Source/RewardGeneratorBasedTMagic.cs
@@ -43,7 +43,7 @@ namespace Flavor_Expansion
                         Thing thing = ThingMaker.MakeThing(enumerable.ToList().RandomElement());
                         if (thing.Label.Contains("Glitter"))
                         {
-                            thing.stackCount = (int)Mathf.Clamp(MedicineStackRange.RandomInRange, 1, 2);
+                            thing.stackCount = Mathf.Clamp(MedicineStackRange.RandomInRange, 1, 2);
                             collectiveMarketValue += thing.stackCount * 100;
                         }
                         else thing.stackCount = MedicineStackRange.RandomInRange;
@@ -51,7 +51,6 @@ namespace Flavor_Expansion
                             continue;
                         outThings.Add(thing);
                         collectiveMarketValue += thing.MarketValue * thing.stackCount;
-                        //Log.Warning("Medicine  " + thing.Label + "  " + thing.MarketValue * thing.stackCount + ", coll  " + collectiveMarketValue);
                     }
                 
                 }
